Fixed category dictionary
A word in the whole dataset but not in the category is now represented in the dictionary as cat[word] = 0. This is because in the algorithm adds 1 to the frequency of a word in category. This was more for csv output purposes as words that don't exist in a dictionary already output a frequency of 0.